1) 플레이어 스텟 매니저 2) 버튼 입력에 따른 스탯 상승 3) 버튼 꾹 누르면 스탯 계속 상승(코루틴 이용) 4) 스탯 상승하면 정보창에 반영 5) 메뉴버튼 클릭에 따라 정보메뉴 바뀌도록

+++ Utility/DataStruct.cs
@@ -115,7 +115,7 @@ public class LivingData : BaseData
 {
     public int level;
     public int EXP;
-    public int damage;
+    public long damage;
     public int hitPoint;
     public int armor;
     public float attackSpeed;
